+++ index.js
@@ -45,7 +45,6 @@ const server = Hapi.server({
 
 const plugins = [
   require('./src/lib/message-queue').plugin,
-  require('./src/modules/billing/register-subscribers'),
   require('./src/modules/returns/register-subscribers'),
   require('./src/plugins/internal-calling-user'),
   require('./src/modules/address-search/plugin')
