--- src/modules/billing/bull-jobs/lib/batch-status.js
@@ -1,20 +1,0 @@
-'use strict';
-
-const { partialRight } = require('lodash');
-
-const Batch = require('../../../../lib/models/batch');
-
-/**
- * Throws an error if the batch is not the provided status
- * @param {Batch} batch
- * @param {String} status
- */
-const assertBatchIsStatus = (batch, status) => {
-  if (!batch.statusIsOneOf(status)) {
-    throw new Error(`Expected ${status} batch status`);
-  }
-};
-
-const assertBatchIsProcessing = partialRight(assertBatchIsStatus, Batch.BATCH_STATUS.processing);
-
-exports.assertBatchIsProcessing = assertBatchIsProcessing;
