+++ src/modules/billing/jobs/populate-batch-charge-versions.js
@@ -1,23 +1,30 @@
 'use strict';
 
-const { get } = require('lodash');
+const { partial, get } = require('lodash');
 
-const batchJob = require('./lib/batch-job');
+const ioRedis = require('../../../lib/connectors/io-redis');
+const connection = ioRedis.createConnection();
+
+// Bull queue setup
+const { Queue, Worker } = require('bullmq');
+const JOB_NAME = 'billing.populate-batch-charge-versions';
+const queue = new Queue(JOB_NAME, { connection });
+
 const batchService = require('../services/batch-service');
 const chargeVersionService = require('../services/charge-version-service');
-const { BATCH_ERROR_CODE } = require('../../../lib/models/batch');
 
-const JOB_NAME = 'billing.populate-batch-charge-versions.*';
+const { BATCH_ERROR_CODE, BATCH_STATUS, BATCH_TYPE } = require('../../../lib/models/batch');
+const helpers = require('./lib/helpers');
+const batchJob = require('./lib/batch-job');
 
-const createMessage = (eventId, batch) => {
-  return batchJob.createMessage(JOB_NAME, batch, { eventId }, {
-    singletonKey: batch.id
-  });
-};
+const processChargeVersionsJob = require('./process-charge-versions');
+const twoPartTariffMatchingJob = require('./two-part-tariff-matching');
 
-const handlePopulateBatch = async job => {
+const createMessage = partial(helpers.createMessage, JOB_NAME);
+
+const handler = async job => {
   batchJob.logHandling(job);
-  const batchId = get(job, 'data.batch.id');
+  const batchId = get(job, 'data.batchId');
 
   try {
     const batch = await batchService.getBatchById(batchId);
@@ -25,13 +32,48 @@ const handlePopulateBatch = async job => {
     // Populate water.billing_batch_charge_version_years
     const billingBatchChargeVersionYears = await chargeVersionService.createForBatch(batch);
 
-    return { billingBatchChargeVersionYears, batch };
+    // If there is nothing to process, mark empty batch
+    if (billingBatchChargeVersionYears.length === 0) {
+      const updatedBatch = await batchService.setStatus(batchId, BATCH_STATUS.empty);
+      return { batch: updatedBatch };
+    }
+
+    return { batch };
   } catch (err) {
     await batchJob.logHandlingErrorAndSetBatchStatus(job, err, BATCH_ERROR_CODE.failedToPopulateChargeVersions);
     throw err;
   }
 };
 
+const onComplete = async job => {
+  batchJob.logOnComplete(job);
+
+  try {
+    const { batch } = job.returnvalue;
+
+    // Don't do any further processing for empty batch
+    if (batch.status === BATCH_STATUS.empty) {
+      return;
+    }
+
+    if (batch.type === BATCH_TYPE.annual) {
+      await processChargeVersionsJob.queue.add(
+        ...processChargeVersionsJob.createMessage(batch)
+      );
+    } else {
+      // Two-part tariff matching for TPT/supplementary run
+      await twoPartTariffMatchingJob.queue.add(
+        ...twoPartTariffMatchingJob.createMessage(batch)
+      );
+    }
+  } catch (err) {
+    batchJob.logOnCompleteError(job);
+  }
+};
+
+const worker = new Worker(JOB_NAME, handler, { connection });
+worker.on('completed', onComplete);
+worker.on('failed', helpers.onFailedHandler);
+
 exports.createMessage = createMessage;
-exports.handler = handlePopulateBatch;
-exports.jobName = JOB_NAME;
+exports.queue = queue;
